# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "go-clean-architecture",
        "title": "go-clean-architecture",
        "contact": {},
        "version": "0.0.1"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/api/article": {
            "post": {
                "description": "Article 을 생성합니다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Article"
                ],
                "summary": "Article 생성",
                "operationId": "create-article",
                "parameters": [
                    {
                        "description": "새로운 Article 정보",
                        "name": "article",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Article"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/domain.Article"
                        }
                    }
                }
            }
        },
        "/api/article/{title}": {
            "get": {
                "description": "단일 Article 을 조회합니다.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Article"
                ],
                "summary": "단일 Article 조회",
                "operationId": "get-article",
                "parameters": [
                    {
                        "type": "string",
                        "description": "조회할 Article 제목",
                        "name": "title",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Article"
                        }
                    }
                }
            },
            "put": {
                "description": "Article 을 수정합니다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Article"
                ],
                "summary": "Article 수정",
                "operationId": "update-article",
                "parameters": [
                    {
                        "type": "string",
                        "description": "수정할 Article 제목",
                        "name": "title",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "수정할 Article 정보",
                        "name": "article",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Article"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Article 을 제거합니다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Article"
                ],
                "summary": "Article 제거",
                "operationId": "delete-article",
                "parameters": [
                    {
                        "type": "string",
                        "description": "삭제할 Article 제목",
                        "name": "title",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/articles": {
            "get": {
                "description": "Article 리스트를 조회합니다.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Article"
                ],
                "summary": "Article 리스트 조회",
                "operationId": "get-articles",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Article"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.Article": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userName": {
                    "type": "string"
                }
            }
        }
    }
}